refactor: relocate productcontext.jsx and filtercontext.jsx to contexts folder
